# artifacts/api-server/src/routes/storage.ts
import { Router, type IRouter, type Request, type Response } from "express";
import { Readable } from "stream";
import {
  RequestUploadUrlBody,
  RequestUploadUrlResponse,
} from "@workspace/api-zod";
import { ObjectStorageService, ObjectNotFoundError } from "../lib/objectStorage";

const router: IRouter = Router();
const objectStorageService = new ObjectStorageService();

/**
 * Organizations allowed to upload media. Mirrors the write-gating used by the
 * spaces catalog (only C2 LABS can edit; OPINNO/AURORA360 are read-only).
 */
const ORG_PERMISSIONS: Record<string, { canEdit: boolean }> = {
  "C2 LABS": { canEdit: true },
  "OPINNO": { canEdit: false },
  "AURORA360": { canEdit: false },
};

/** Accepted upload content-type prefixes. */
const ALLOWED_CONTENT_TYPE_PREFIXES = ["image/", "video/"];
/** Maximum accepted upload size (200 MB), enforced server-side. */
const MAX_UPLOAD_BYTES = 200 * 1024 * 1024;

/**
 * POST /storage/uploads/request-url
 *
 * Request a presigned URL for file upload.
 * The client sends JSON metadata (name, size, contentType) — NOT the file.
 * Then uploads the file directly to the returned presigned URL.
 *
 * Gated to editing organizations (C2 LABS) and validated for media type/size.
 */
router.post("/storage/uploads/request-url", async (req: Request, res: Response) => {
  const session = req.session as any;
  const userOrg = session?.userOrg || "";
  if (!session?.userEmail) {
    res.status(401).json({ error: "Authentication required" });
    return;
  }
  const perms = ORG_PERMISSIONS[userOrg] || { canEdit: false };
  if (!perms.canEdit) {
    res.status(403).json({ error: "You do not have permission to upload media" });
    return;
  }

  const parsed = RequestUploadUrlBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: "Missing or invalid required fields" });
    return;
  }

  const { name, size, contentType } = parsed.data;

  const type = (contentType || "").toLowerCase();
  if (!ALLOWED_CONTENT_TYPE_PREFIXES.some(prefix => type.startsWith(prefix))) {
    res.status(400).json({ error: "Only image and video files are allowed" });
    return;
  }
  if (typeof size === "number" && size > MAX_UPLOAD_BYTES) {
    res.status(400).json({ error: "File too large (max 200 MB)" });
    return;
  }

  try {
    const uploadURL = await objectStorageService.getObjectEntityUploadURL();
    const objectPath = objectStorageService.normalizeObjectEntityPath(uploadURL);

    res.json(
      RequestUploadUrlResponse.parse({
        uploadURL,
        objectPath,
        metadata: { name, size, contentType },
      }),
    );
  } catch (error) {
    req.log.error({ err: error }, "Error generating upload URL");
    res.status(500).json({ error: "Failed to generate upload URL" });
  }
});

/**
 * GET /storage/public-objects/*
 *
 * Serve public assets from PUBLIC_OBJECT_SEARCH_PATHS.
 * These are unconditionally public — no authentication or ACL checks.
 * IMPORTANT: Always provide this endpoint when object storage is set up.
 */
router.get("/storage/public-objects/*filePath", async (req: Request, res: Response) => {
  try {
    const raw = req.params.filePath;
    const filePath = Array.isArray(raw) ? raw.join("/") : raw;
    const file = await objectStorageService.searchPublicObject(filePath);
    if (!file) {
      res.status(404).json({ error: "File not found" });
      return;
    }

    const response = await objectStorageService.downloadObject(file);

    res.status(response.status);
    response.headers.forEach((value, key) => res.setHeader(key, value));

    if (response.body) {
      const nodeStream = Readable.fromWeb(response.body as ReadableStream<Uint8Array>);
      nodeStream.pipe(res);
    } else {
      res.end();
    }
  } catch (error) {
    req.log.error({ err: error }, "Error serving public object");
    res.status(500).json({ error: "Failed to serve public object" });
  }
});

/**
 * GET /storage/objects/*
 *
 * Serve object entities from PRIVATE_OBJECT_DIR.
 * Served from a separate path from /public-objects. Access requires an
 * authenticated session — any logged-in organizer (C2 LABS / OPINNO / AURORA360)
 * may view uploaded space media; anonymous requests are rejected.
 */
router.get("/storage/objects/*path", async (req: Request, res: Response) => {
  const session = req.session as any;
  if (!session?.userEmail) {
    res.status(401).json({ error: "Authentication required" });
    return;
  }
  try {
    const raw = req.params.path;
    const wildcardPath = Array.isArray(raw) ? raw.join("/") : raw;
    const objectPath = `/objects/${wildcardPath}`;
    const objectFile = await objectStorageService.getObjectEntityFile(objectPath);

    const response = await objectStorageService.downloadObject(objectFile);

    res.status(response.status);
    response.headers.forEach((value, key) => res.setHeader(key, value));

    if (response.body) {
      const nodeStream = Readable.fromWeb(response.body as ReadableStream<Uint8Array>);
      nodeStream.pipe(res);
    } else {
      res.end();
    }
  } catch (error) {
    if (error instanceof ObjectNotFoundError) {
      req.log.warn({ err: error }, "Object not found");
      res.status(404).json({ error: "Object not found" });
      return;
    }
    req.log.error({ err: error }, "Error serving object");
    res.status(500).json({ error: "Failed to serve object" });
  }
});

export default router;